requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Werkzeug
python-dotenv
Flask-Migrate

Flask-Mail

PyMySQL
